requirements1.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-openai
langchain-huggingface
faiss-cpu

